added owner-cheching for methods: findOne(), delete() (there are not exception handlers for these methods: UnauthorizedEsxception). Edited search-method -> stream receive current user's list of aspects, not list of aspects by ALL users.

+++ src/main/java/com/sgkhmjaes/jdias/service/impl/AspectServiceImpl.java
@@ -1,14 +1,23 @@
 package com.sgkhmjaes.jdias.service.impl;
 
+import com.google.common.collect.Lists;
+import com.sgkhmjaes.jdias.domain.Person;
+import com.sgkhmjaes.jdias.domain.UserAccount;
+import com.sgkhmjaes.jdias.repository.PersonRepository;
+import com.sgkhmjaes.jdias.repository.UserAccountRepository;
+import com.sgkhmjaes.jdias.repository.UserRepository;
+import com.sgkhmjaes.jdias.security.SecurityUtils;
 import com.sgkhmjaes.jdias.service.AspectService;
 import com.sgkhmjaes.jdias.domain.Aspect;
 import com.sgkhmjaes.jdias.repository.AspectRepository;
 import com.sgkhmjaes.jdias.repository.search.AspectSearchRepository;
+import com.sgkhmjaes.jdias.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
@@ -25,15 +34,18 @@ public class AspectServiceImpl implements AspectService{
     private final Logger log = LoggerFactory.getLogger(AspectServiceImpl.class);
 
     private final AspectRepository aspectRepository;
-
     private final AspectSearchRepository aspectSearchRepository;
+    private final UserService userService;
 
-    public AspectServiceImpl(AspectRepository aspectRepository, AspectSearchRepository aspectSearchRepository) {
+    public AspectServiceImpl(AspectRepository aspectRepository,
+                             AspectSearchRepository aspectSearchRepository,
+                             UserService userService) {
         this.aspectRepository = aspectRepository;
         this.aspectSearchRepository = aspectSearchRepository;
+        this.userService = userService;
     }
 
-    /**
+    /**se
      * Save a aspect.
      *
      * @param aspect the entity to save
@@ -47,16 +59,22 @@ public class AspectServiceImpl implements AspectService{
         return result;
     }
 
-    /**
+    /*/**
      *  Get all the aspects.
      *
      *  @return the list of entities
-     */
+     *
     @Override
     @Transactional(readOnly = true)
     public List<Aspect> findAll() {
         log.debug("Request to get all Aspects");
         return aspectRepository.findAll();
+    }*/
+
+    @Override
+    public List<Aspect> findAllByUser() {
+        log.debug("Request to get all Aspects by user");
+        return new ArrayList<>(userService.getCurrentPerson().getAspects());
     }
 
     /**
@@ -69,7 +87,12 @@ public class AspectServiceImpl implements AspectService{
     @Transactional(readOnly = true)
     public Aspect findOne(Long id) {
         log.debug("Request to get Aspect : {}", id);
-        return aspectRepository.findOne(id);
+        Aspect foundAspect =  aspectRepository.findOne(id);
+
+        if(userService.getCurrentPerson().getAspects().contains(foundAspect))
+            return foundAspect;
+        else return null;
+        // or return new UnauthorizedException
     }
 
     /**
@@ -80,8 +103,12 @@ public class AspectServiceImpl implements AspectService{
     @Override
     public void delete(Long id) {
         log.debug("Request to delete Aspect : {}", id);
-        aspectRepository.delete(id);
-        aspectSearchRepository.delete(id);
+        Aspect foundAspect =  aspectRepository.findOne(id);
+
+        if(userService.getCurrentPerson().getAspects().contains(foundAspect)) {
+            aspectRepository.delete(id);
+            aspectSearchRepository.delete(id);
+        }
     }
 
     /**
@@ -94,8 +121,10 @@ public class AspectServiceImpl implements AspectService{
     @Transactional(readOnly = true)
     public List<Aspect> search(String query) {
         log.debug("Request to search Aspects for query {}", query);
-        return StreamSupport
-            .stream(aspectSearchRepository.search(queryStringQuery(query)).spliterator(), false)
+
+        Iterable<Aspect> userAspects = new ArrayList<>(userService.getCurrentPerson().getAspects());
+
+        return StreamSupport.stream(userAspects.spliterator(), false)
             .collect(Collectors.toList());
     }
 }
